+++ app/infrastructure/resilience/service.py
@@ -3,6 +3,7 @@
 Provides a class-based interface to circuit breakers and retry stores for easier DI and testing.
 """
 
+from functools import cache
 from typing import Any, Callable, Dict, Optional, TYPE_CHECKING
 
 import structlog
@@ -12,6 +13,7 @@ from infrastructure.resilience.circuit_breaker import (
     CircuitBreaker,
     CircuitState,
 )
+from infrastructure.configuration.infrastructure.retry import get_retry_settings
 
 if TYPE_CHECKING:
     from infrastructure.resilience.retry.store import RetryStore
@@ -237,3 +239,27 @@ class ResilienceService:
             List of circuit breaker names
         """
         return list(self._circuit_breakers.keys())
+
+
+@cache
+def get_resilience_service() -> ResilienceService:
+    """Get application-scoped resilience service singleton.
+
+    Returns a ResilienceService instance that provides unified access to
+    circuit breakers and retry stores for fault-tolerant operations.
+
+    Usage:
+        from infrastructure.services import ResilienceServiceDep
+
+        @router.get("/external-call")
+        def make_call(resilience: ResilienceServiceDep):
+            cb = resilience.get_or_create_circuit_breaker("external_api")
+            try:
+                return cb.call(external_api_function)
+            except CircuitBreakerOpenError:
+                return {"error": "Service unavailable"}
+
+    Returns:
+        ResilienceService: Cached resilience service instance
+    """
+    return ResilienceService(retry_settings=get_retry_settings())
